app/src/main/java/com/lanshiqin/algorithm/leetcode/a21/Solution.java
package com.lanshiqin.algorithm.leetcode.a21;

import com.lanshiqin.algorithm.leetcode.ListNode;

/**
 * 21. 合并两个有序链表
 * 将两个升序链表合并为一个新的 升序 链表并返回。新链表是通过拼接给定的两个链表的所有节点组成的。
 * <p>
 * <p>
 * 示例 1：
 * <p>
 * 输入：l1 = [1,2,4], l2 = [1,3,4]
 * 输出：[1,1,2,3,4,4]
 * <p>
 * <p>
 * 示例 2：
 * <p>
 * 输入：l1 = [], l2 = []
 * 输出：[]
 * <p>
 * <p>
 * 示例 3：
 * <p>
 * 输入：l1 = [], l2 = [0]
 * 输出：[0]
 * <p>
 * 提示：
 * <p>
 * 两个链表的节点数目范围是 [0, 50]
 * -100 <= Node.val <= 100
 * l1 和 l2 均按 非递减顺序 排列
 * <p>
 * 来源：力扣（LeetCode）
 * 链接：https://leetcode-cn.com/problems/merge-two-sorted-lists
 * 著作权归领扣网络所有。商业转载请联系官方授权，非商业转载请注明出处。
 */
public class Solution {

    public ListNode mergeTwoLists(ListNode l1, ListNode l2) {
        if (l1 == null && l2 == null) return null;
        ListNode root = new ListNode();
        ListNode cursor = root;
        while (l1 != null && l2 != null) {
            if (l1.val > l2.val) {
                cursor.next = new ListNode(l2.val);
                l2 = l2.next;
            } else {
                cursor.next = new ListNode(l1.val);
                l1 = l1.next;
            }
            cursor = cursor.next;
        }
        cursor.next = l1 == null ? l2 : l1;
        return root.next;
    }
}
